requirements.txt
python-telegram-bot==21.5
dateparser==1.2.0
APScheduler==3.10.4
SQLAlchemy==2.0.31
psycopg2-binary==2.9.9
google-api-python-client==2.139.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
python-dotenv==1.0.1
pytz==2024.1
